requirements.txt
torch>=2.2.0
torchaudio
openai-whisper
TTS==0.22.0
numpy==1.22.0
simpleaudio==1.0.4
sounddevice
transformers
